# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Galaktika galaxy game",
        "title": "Galaktika API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/ship-models": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ship-models"
                ],
                "summary": "List all ship models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/galaxy.ShipModel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ship-models"
                ],
                "summary": "Create a ship model",
                "parameters": [
                    {
                        "description": "ShipModel data",
                        "name": "shipModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ship-models/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ship-models"
                ],
                "summary": "Get a ship model by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ShipModel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ship-models"
                ],
                "summary": "Update a ship model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ShipModel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ShipModel data",
                        "name": "shipModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ship-models"
                ],
                "summary": "Delete a ship model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ShipModel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "galaxy.ShipModel": {
            "type": "object",
            "properties": {
                "cargo_mass": {
                    "type": "number"
                },
                "defense_mass": {
                    "type": "number"
                },
                "engine_mass": {
                    "type": "number"
                },
                "guns": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "one_gun_mass": {
                    "type": "number"
                },
                "owner_id": {
                    "type": "string"
                }
            }
        },
...
    }
}